tried to separate web service from main file
